Dependencies/requirements.txt
python_dateutil==2.8.2
pywin32==305
selenium==4.8.3
webdriver_manager==3.8.5
